refactor: search 리팩토링
